requirements.txt
azure-functions
azure-functions-worker
agrobr
geopandas
shapely
